auto-refactoring

+++ src/users/users.service.ts
@@ -11,55 +11,86 @@ export class UsersService {
     @InjectRepository(User) private UserRepository: Repository<User>,
   ) {}
 
-  async getAll(role?: string): Promise<UserResponseDto[]> {
-    let users =  await this.UserRepository.find();
+  async getAll(
+    role?: string,
+    fullName?: string,
+    efficiency?: number,
+  ): Promise<UserResponseDto[]> {
+    const query = this.UserRepository.createQueryBuilder('user');
     if (role) {
-      users = role ? users.filter(user => user.role === role) : users;
+      query.andWhere('user.role = :role', { role });
     }
+    if (fullName) {
+      query.andWhere('user.full_name = :fullName', { fullName });
+    }
+    if (efficiency !== undefined) {
+      query.andWhere('user.efficiency = :efficiency', { efficiency });
+    }
+    const users = await query.getMany();
     return users;
   }
 
   async getOneById(id: number): Promise<UserResponseDto | undefined> {
-      return await this.UserRepository.findOne({
-        where: { id: id },
-      });
+    return await this.UserRepository.findOne({
+      where: { id: id },
+    });
   }
 
-  async create(user: CreateUserDto): Promise<User | undefined> {
-    let existName = await this.UserRepository.findOneBy({
+  async create(user: CreateUserDto): Promise<User> {
+    const existName = await this.UserRepository.findOneBy({
       full_name: user.full_name,
     });
     if (existName) {
-      return undefined;
+      throw new HttpException(
+        {
+          success: false,
+          result: { error: 'user with this full_name already exist' },
+        },
+        HttpStatus.CONFLICT,
+      );
     }
     const createdUser = this.UserRepository.create(user);
     return await this.UserRepository.save(createdUser);
   }
 
-  async update(id: number, user: UpdateUserDto):Promise<UserResponseDto | undefined> {
+  async update(id: number, user: UpdateUserDto): Promise<UserResponseDto> {
     let foundUser = await this.UserRepository.findOneBy({
       id: id,
     });
-    let existName = await this.UserRepository.findOneBy({
+    const existName = await this.UserRepository.findOneBy({
       full_name: user.full_name,
     });
     if (existName) {
-      return undefined;
+      throw new HttpException(
+        {
+          success: false,
+          result: { error: 'user with this full_name already exist' },
+        },
+        HttpStatus.CONFLICT,
+      );
     }
     foundUser = { ...foundUser, ...user, updated_at: new Date() };
     return await this.UserRepository.save(foundUser);
   }
 
-  async delete(id?: number): Promise<UserResponseDto | undefined> {
-    if(id) {
-      let foundUser = await this.UserRepository.findOneBy({
+  async delete(id?: number): Promise<UserResponseDto> {
+    if (id) {
+      const foundUser = await this.UserRepository.findOneBy({
         id: id,
       });
+      if (!foundUser) {
+        throw new HttpException(
+          {
+            success: false,
+            result: { error: 'user with this id not found' },
+          },
+          HttpStatus.NOT_FOUND,
+        );
+      }
       await this.UserRepository.delete(id);
       return foundUser;
     } else {
-      await this.UserRepository.clear()
-      return undefined;
+      await this.UserRepository.clear();
     }
   }
 }
